requirements.txt
apache-airflow
tweepy
boto3
python-dotenv
pandas